Created a new `MetadataQuery` class to reduce duplicate code
Moving away from all queries setting metadata themselves, and instead having a `MetadataQuery` that can be set in each query that supports metadata
